requirements.txt
django
djangorestframework
weasyprint
pandas
django-multiselectfield

